requirements.txt

# pip install -i https://pypi.tuna.tsinghua.edu.cn/simple
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)

ipdb
matplotlib
gdown
einops
joblib
open3d
transforms3d
pandas
mmcv
ruamel.yaml
wandb
trimesh
timm
pypng
diffusers
accelerate
pin
objaverse
PyOpenGL
PyOpenGL_accelerate
vispy
pycocotools
xformers
easydict
seaborn
torchgeometry
huggingface_hub==0.34.4
roma
pillow-heif
pyglet<2
celery
panda3d
webdataset
lpips
torch-cluster
PyMCubes
plotly
torch_scatter
peft
torchmetrics[image]

# MAX_JOBS=4 pip install flash-attn==2.7.4.post1 --no-build-isolation
git+https://github.com/lessw2020/Ranger-Deep-Learning-Optimizer
git+https://github.com/facebookresearch/pytorch3d.git
git+https://github.com/PuchatekwSzortach/imgaug.git@numpy_fix
git+https://github.com/facebookresearch/detectron2.git
git+https://github.com/zuodexin/megapose6d.git
git+https://github.com/princeton-vl/lietorch.git
git+https://github.com/pvigier/perlin-numpy
git+https://github.com/facebookresearch/segment-anything.git